+++ packages/battleship/src/random_generator_trait.rs
@@ -2,5 +2,5 @@ use mockall::automock;
 
 #[automock]
 pub trait RandomGeneratorTrait {
-    fn generate(&mut self, start: u32, end: u32) -> u32;
+    fn generate(&mut self, start: u8, end: u8) -> u8;
 }
